gobject: Fix CMake references to libinsane
The CMakeLists.txt in libinsane-gobject relies heavily on a specific
directory structure and also uses a few workarounds that reference
relative paths that might not be there when using our new top-level
CMakeLists.txt.

However, I've added a few workarounds as well, because right now we want
to still support the Makefile that resides in the top-level until the
buildbot scripts were adjusted so that they build from the top-level
CMakeLists.txt.

So for now this links the gobject bindings statically against libinsane,
which we don't want to necessarily have in the long term (only optional
at least).

Signed-off-by: aszlig <aszlig@nix.build>